More verification tests
