+++ framework/src/play-integration-test/src/test/scala/play/it/http/AkkaResponseHeaderHandlingSpec.scala
@@ -6,7 +6,7 @@ package play.it.http
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.mvc._
 import play.api.test.{ PlaySpecification, Port }
-import play.it.AkkaHttpIntegrationSpecification
+import play.it.{ AkkaHttpIntegrationSpecification, LogTester }
 
 class AkkaResponseHeaderHandlingSpec extends PlaySpecification with AkkaHttpIntegrationSpecification {
 
@@ -45,6 +45,22 @@ class AkkaResponseHeaderHandlingSpec extends PlaySpecification with AkkaHttpInte
       )
       responses(0).headers.get("Link") must_== Some("""<http://example.com/some/url>; rel=next""")
       //responses(0).headers.get("Link") must_== Some("""<http://example.com/some/url>; rel="next"""")
+    }
+
+    "don't log a warning for Set-Cookie headers with negative ages" in {
+      val problemHeaderValue = "PLAY_FLASH=; Max-Age=-86400; Expires=Tue, 30 Jan 2018 06:29:53 GMT; Path=/; HTTPOnly"
+      withServer((Action, _) => Action { rh =>
+        // Test the header reported in https://github.com/playframework/playframework/issues/8205
+        Results.Ok.withHeaders("Set-Cookie" -> problemHeaderValue)
+      }) { port =>
+        val (Seq(response), logMessages) = LogTester.recordLogEvents {
+          BasicHttpClient.makeRequests(port)(
+            BasicRequest("GET", "/", "HTTP/1.1", Map(), "")
+          )
+        }
+        response.status must_== 200
+        logMessages.map(_.getFormattedMessage) must not contain (contain(problemHeaderValue))
+      }
     }
 
   }
